requirements.txt
amqp==2.5.0
aniso8601==6.0.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dict2xml==1.5
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT==0.3.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
idna==2.8
inflect==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
json2xml==3.0.1
jsonschema==3.0.1
kombu==4.6.3
MarkupSafe==1.1.1
pika==1.1.0
PyJWT==1.4.2
PyMySQL==0.9.3
pyrsistent==0.15.2
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlacodegen==2.0.1
SQLAlchemy==1.3.2
toml==0.10.0
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.2
xmltodict==0.11.0
pycryptodomex
pyOpenSSL==19.0.0
